✨ feat(server): 添加push接口

+++ models/server.go
@@ -3,7 +3,7 @@ package models
 // ServerUniversalQueryReq 节点访问server相关接口通用query参数
 type ServerUniversalQueryReq struct {
 	// 节点ID
-	NodeId int `form:"node_id" binding:"required"`
+	NodeId int64 `form:"node_id" binding:"required"`
 	// 节点类型
 	NodeType string `form:"node_type" binding:"oneof=v2ray trojan"`
 	// 认证token
